@@ -1,17 +1,14 @@
-import React, {FC, useState} from 'react';
-import {useTheme} from '@material-ui/core/styles';
-import MobileStepper from '@material-ui/core/MobileStepper';
-import Button from '@material-ui/core/Button';
-import KeyboardArrowLeft from '@material-ui/icons/KeyboardArrowLeft';
-import KeyboardArrowRight from '@material-ui/icons/KeyboardArrowRight';
+import {FC, useState} from 'react';
+import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
+import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
 import BeerForm from "./Form";
-import {Grid} from '@material-ui/core';
-import "./BeerFormStepper.scss";
+import {Button, Grid, MobileStepper} from '@mui/material';
 import {useRoomContext} from "../../hooks/useContextHook";
+import "./BeerFormStepper.scss";
+
 
 const BeerFormStepper: FC = () => {
   const {beers} = useRoomContext();
-  const theme = useTheme();
 
   const [activeStep, setActiveStep] = useState<number>(0);
   const maxSteps = beers.length;
@@ -23,7 +20,7 @@ const BeerFormStepper: FC = () => {
   if (maxSteps === 0) return null;
 
   return (
-    <div className='beer-form-stepper-root'>
+    <div className="beer-form-stepper-root">
       <MobileStepper
         steps={maxSteps}
         position="static"
@@ -32,12 +29,12 @@ const BeerFormStepper: FC = () => {
         nextButton={
           <Button size="small" onClick={handleNext} disabled={activeStep === maxSteps - 1}>
             Następne
-            {theme.direction === 'rtl' ? <KeyboardArrowLeft/> : <KeyboardArrowRight/>}
+            <KeyboardArrowRight/>
           </Button>
         }
         backButton={
           <Button size="small" onClick={handleBack} disabled={activeStep === 0}>
-            {theme.direction === 'rtl' ? <KeyboardArrowRight/> : <KeyboardArrowLeft/>}
+            <KeyboardArrowLeft/>
             Poprzednie
           </Button>
         }
@@ -65,6 +62,6 @@ const BeerFormStepper: FC = () => {
       />
     </div>
   );
-}
+};
 
 export default BeerFormStepper;
