@@ -1,6 +1,6 @@
 import React, {FC} from "react";
 import {useRoomContext} from "../../hooks/useContextHook";
-import {List, ListItem, ListItemText} from "@material-ui/core";
+import {List, ListItem, ListItemText} from "@mui/material";
 
 
 const Participants: FC = () => {
