@@ -1,10 +1,5 @@
-import React, {FC} from "react";
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
+import {FC} from "react";
+import {Paper, Table, TableBody, TableContainer, TableHead, TableRow} from '@mui/material';
 import {useRoomContext} from "../../hooks/useContextHook";
 import {
   TableHeaderCell,
@@ -48,6 +43,6 @@ const UserRatingsTable: FC = () => {
       </Table>
     </TableContainer>
   );
-}
+};
 
-export default UserRatingsTable
+export default UserRatingsTable;
