@@ -1,4 +1,4 @@
-import React, {FC} from "react";
+import {FC} from "react";
 import {BrowserRouter, Route, Switch} from "react-router-dom";
 import Home from "../components/pages/Home";
 import Room from "../components/pages/Room";
